Fix spelling of Sudoku throughout.
Other minor tweaks.

+++ src/HUD.cpp
@@ -89,8 +89,8 @@ void HUD::handleMouseDown(const int &x, const int &y, const MouseButtonState& bu
     int max_z_index = -INT_MAX;
     for (auto &i : stack) {
         const glm::ivec2 bottomLeft = glm::ivec2(static_cast<glm::vec3*>(i->data)[1].x, static_cast<glm::vec3*>(i->data)[1].y);
-        if (x >= bottomLeft.x && x < bottomLeft.x + i->overlay->getWidth() &&
-            y >= bottomLeft.y && y < bottomLeft.y + i->overlay->getHeight() ) {
+        if (x >= bottomLeft.x && x < bottomLeft.x + static_cast<int>(i->overlay->getWidth()) &&
+            y >= bottomLeft.y && y < bottomLeft.y + static_cast<int>(i->overlay->getHeight())) {
             if (i->zIndex > max_z_index) {
                 intersected_item = i;
                 max_z_index = i->zIndex;
@@ -113,8 +113,8 @@ void HUD::handleMouseUp(const int &x, const int &y, const MouseButtonState& butt
     int max_z_index = -INT_MAX;
     for (auto &i : stack) {
         const glm::ivec2 bottomLeft = glm::ivec2(static_cast<glm::vec3*>(i->data)[1].x, static_cast<glm::vec3*>(i->data)[1].y);
-        if (x >= bottomLeft.x && x < bottomLeft.x + i->overlay->getWidth() &&
-            y >= bottomLeft.y && y < bottomLeft.y + i->overlay->getHeight() ) {
+        if (x >= bottomLeft.x && x < bottomLeft.x + static_cast<int>(i->overlay->getWidth()) &&
+            y >= bottomLeft.y && y < bottomLeft.y + static_cast<int>(i->overlay->getHeight())) {
             if (i->zIndex > max_z_index) {
                 intersected_item = i;
                 max_z_index = i->zIndex;
@@ -137,8 +137,8 @@ void HUD::handleMouseDrag(const int &x, const int &y, const MouseButtonState& bu
     int max_z_index = -INT_MAX;
     for (auto &i : stack) {
         const glm::ivec2 bottomLeft = glm::ivec2(static_cast<glm::vec3*>(i->data)[1].x, static_cast<glm::vec3*>(i->data)[1].y);
-        if (x >= bottomLeft.x && x < bottomLeft.x + i->overlay->getWidth() &&
-            y >= bottomLeft.y && y < bottomLeft.y + i->overlay->getHeight() ) {
+        if (x >= bottomLeft.x && x < bottomLeft.x + static_cast<int>(i->overlay->getWidth()) &&
+            y >= bottomLeft.y && y < bottomLeft.y + static_cast<int>(i->overlay->getHeight())) {
             if (i->zIndex > max_z_index) {
                 intersected_item = i;
                 max_z_index = i->zIndex;
